MVP tg бота

+++ internal/domain/user.go
@@ -3,5 +3,5 @@ package domain
 type User struct {
 	ID         int    `json:"id"`
 	Name       string `json:"name" validate:"required,min=2"`
-	TelegramID int    `json:"telegram_id"`
+	TelegramID int64  `json:"telegram_id"`
 }
